+++ CHANGES.rst
@@ -21,7 +21,7 @@ New Features
   - ``angular_diameter_distance_z1z2`` now supports the computation of
     the angular diameter distance between a scalar and an array like
     argument. [#4593]
-  
+
 - ``astropy.io.ascii``
 
   - File name could be passed as ``Path`` object. [#4606]
@@ -415,6 +415,8 @@ Bug Fixes
 - ``astropy.nddata``
 
 - ``astropy.stats``
+
+  - Fixed wrong results when using median_absolute_deviation with masked arrays
 
 - ``astropy.table``